+++ experiments/src/best-blocks.ts
@@ -1,4 +1,4 @@
-import { getObservableClient } from "@polkadot-api/client"
+import { getObservableClient } from "@polkadot-api/observable-client"
 import { getScProvider } from "@polkadot-api/sc-provider"
 import { WellKnownChain, createScClient } from "@substrate/connect"
 import { createClient } from "@polkadot-api/substrate-client"