+++ packages/client/src/index.ts
@@ -1,7 +1,5 @@
 export * from "./client"
-export * from "./observableClient"
 export * from "./re-exports"
-export type { BlockInfo } from "./observableClient"
 export type { EventPhase } from "./event"
 export type { PolkadotSigner } from "@polkadot-api/polkadot-signer"
 export type { PolkadotClient, TypedApi, FixedSizeArray } from "./types"