+++ packages/client/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Moved `getObservableClient` and all derived types into a separate `@polkadot-api/observable-client` package.
+
 ## 0.0.1 - 2024-04-03
 
 ### Changed